refactor(node): move some global node options to their respective node modules
